+++ CHANGELOG.md
@@ -5,7 +5,7 @@ We follow [Semantic Versions](https://semver.org/).
 ## Unreleased
 
 - Confirm support for python 3.14
-- Add `OrderingFilterBackendWithSecondarySorting` filter
+- Add secondary sorting feature for getting deterministic results in `OrderingFilterBackend`
 
 ## 0.2.2