+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We follow [Semantic Versions](https://semver.org/).
 
 - Confirm support for python 3.14
 - Add secondary sorting feature for getting deterministic results in `OrderingFilterBackend` (and `SARITASA_DRF_ORDERING_ADD_PK_TO_ORDERING` setting)
+- Add `capture_on_commit` to `ApiActionTester`
 
 ## 0.2.2